+++ cmd/agentctl/main.go
@@ -427,6 +427,7 @@ func filterAgentOwners(refs []meta_v1.OwnerReference) (filtered []meta_v1.OwnerR
 func cloudConfigCmd() *cobra.Command {
 	var (
 		stackID string
+		apiURL  string
 		apiKey  string
 	)
 
@@ -448,7 +449,7 @@ config that may be used with this agent.`,
 				return fmt.Errorf("--api-key must be provided")
 			}
 
-			cli := grafanacloud.NewClient(nil, apiKey)
+			cli := grafanacloud.NewClient(nil, apiKey, apiURL)
 
 			ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
 			defer cancel()
@@ -466,6 +467,7 @@ config that may be used with this agent.`,
 
 	cmd.Flags().StringVarP(&stackID, "stack", "u", "", "stack ID to get a config for")
 	cmd.Flags().StringVarP(&apiKey, "api-key", "p", "", "API key to authenticate against Grafana Cloud's API with")
+	cmd.Flags().StringVarP(&apiURL, "api-url", "e", "", "Grafana Cloud's API url")
 	must(cmd.MarkFlagRequired("stack"))
 	must(cmd.MarkFlagRequired("api-key"))